Refactor PageBuilder and Row classes: * Move Row class as static inner class of PageBuilder. * Change class modifier of PageBuilder.Row with private. * Change method modifier of Row#writeXXX with private. * Add comment on Row class definition.
